Read rotational_shifts as an int, not a float
Without this, AugmentedResnetDataLoader (in
augmented_resnet_feature_extractor.py) causes trouble down the line as
its .__len__ method will return a float.

+++ arrows/pytorch/resnet_augmentation.py
@@ -190,7 +190,7 @@ class DataAugmentation(KwiverProcess):
             self._gpu_list = list(map(int, GPU_list_str.split(',')))
 
         # Augmentation variables
-        self._rotational_shifts = float(self.config_value('rotational_shifts'))
+        self._rotational_shifts = int(self.config_value('rotational_shifts'))
         self._resize_factor = float(self.config_value('resize_factor'))
         self._int_shift_factor = float(self.config_value('int_shift_factor'))
